fix(hal): offset addrs by `isize` instead of `i32` (#509)
This changes the `VAddr::offset` and `PAddr::offset` methods to takee an
`isize` offset instead of an `i32` offset. I don't know why I used `i32`
here, that was stupid.

Fixes #507

+++ src/arch/x86_64/interrupt.rs
@@ -48,7 +48,7 @@ static TSS: sync::Lazy<task::StateSegment> = sync::Lazy::new(|| {
     tss.interrupt_stacks[Idt::DOUBLE_FAULT_IST_OFFSET] = unsafe {
         // safety: asdf
         VAddr::from_usize_unchecked(core::ptr::addr_of!(DOUBLE_FAULT_STACK) as usize)
-            .offset(DOUBLE_FAULT_STACK_SIZE as i32)
+            .offset(DOUBLE_FAULT_STACK_SIZE as isize)
     };
     tracing::debug!(?tss, "TSS initialized");
     tss
